requirements.txt
tqdm
opencv-contrib-python-headless
skimage
scikit-image
memorize.py
dlib
PySide2